+++ packages/web-shared/src/components/ui/data-inspector.tsx
@@ -38,6 +38,7 @@ import {
   inspectorThemeExtendedLight,
   inspectorThemeLight,
 } from './inspector-theme';
+import { Button } from './button';
 import { Spinner } from './spinner';
 
 // ---------------------------------------------------------------------------
@@ -152,16 +153,9 @@ function EncryptedInlineLabel() {
   const ctx = useContext(DecryptClickContext);
   if (ctx) {
     return (
-      <button
-        type="button"
-        className="inline-flex items-center gap-1 px-1.5 py-0.5 rounded text-[11px] cursor-pointer"
-        style={{
-          backgroundColor: 'var(--ds-gray-100)',
-          color: 'var(--ds-gray-700)',
-          border: '1px solid var(--ds-gray-400)',
-          fontStyle: 'italic',
-          opacity: ctx.isDecrypting ? 0.6 : 1,
-        }}
+      <Button
+        size="xs"
+        className="align-baseline gap-x-1"
         disabled={ctx.isDecrypting}
         onClick={(e) => {
           e.stopPropagation();
@@ -170,15 +164,12 @@ function EncryptedInlineLabel() {
         title="Click to decrypt"
       >
         {ctx.isDecrypting ? (
-          <Spinner size={12} />
+          <Spinner size={10} />
         ) : (
-          <Lock
-            className="h-3 w-3"
-            style={{ display: 'inline', flexShrink: 0 }}
-          />
+          <Lock className="h-3 w-3" />
         )}
-        <span>{ctx.isDecrypting ? 'Decrypting…' : 'Decrypt'}</span>
-      </button>
+        <span>Decrypt</span>
+      </Button>
     );
   }
   return (
